+++ Tutti/Hints/Models/StandardDeferredHint.swift
@@ -51,7 +51,11 @@ open class StandardDeferredHint: StandardHint, DeferredHint {
     
     open override func present(with presenter: HintPresenter, in vc: UIViewController, from view: UIView) {
         registerPresentationAttempt()
-        guard shouldBePresented else { return }
         super.present(with: presenter, in: vc, from: view)
+    }
+    
+    open override func present(with presenter: HintPresenter, in vc: UIViewController, from item: UIBarButtonItem) {
+        registerPresentationAttempt()
+        super.present(with: presenter, in: vc, from: item)
     }
 }