+++ Tutti/Hints/Presenters/AlertHintPresenter.swift
@@ -23,11 +23,22 @@ public class AlertingHintPresenter: HintPresenter {
     public init() {}
     
     public func present(hint: Hint, in vc: UIViewController, from view: UIView) {
+        alert(hint: hint, from: vc)
+    }
+    
+    public func present(hint: Hint, in vc: UIViewController, from barButtonItem: UIBarButtonItem) {
+        alert(hint: hint, from: vc)
+    }
+    
+    public func dismiss(hint: Hint) { }
+}
+
+private extension AlertingHintPresenter {
+    
+    func alert(hint: Hint, from vc: UIViewController) {
         let alert = UIAlertController(title: hint.title, message: hint.text, preferredStyle: .alert)
         let ok = UIAlertAction(title: "OK", style: .default, handler: nil)
         alert.addAction(ok)
         vc.present(alert, animated: true, completion: nil)
     }
-    
-    public func dismiss(hint: Hint) { }
 }