+++ Tutti/Hints/Presenters/CalloutHintPresenter.swift
@@ -18,7 +18,7 @@
 
 import UIKit
 
-public class CalloutHintPresenter: HintPresenter {
+public class CalloutHintPresenter: HintPresenter, CalloutViewDelegate {
     
     
     // MARK: - Initialization
@@ -26,11 +26,45 @@ public class CalloutHintPresenter: HintPresenter {
     public init() {}
     
     
+    // MARK: - Properties
+    
+    private var presentedCallouts = [CalloutView]()
+    
+    
     // MARK: - Public Functions
 
-    public func present(hint: Hint, in vc: UIViewController, from view: UIView) {
-        CalloutView.show(forView: view, text: hint.text)
+    open func present(hint: Hint, in vc: UIViewController, from view: UIView) {
+        dismiss(hint: hint)
+        createCallout(for: hint).show(forView: view)
+        
     }
     
-    public func dismiss(hint: Hint) { }
+    open func present(hint: Hint, in vc: UIViewController, from item: UIBarButtonItem) {
+        dismiss(hint: hint)
+        createCallout(for: hint).show(forItem: item)
+    }
+    
+    open func dismiss(hint: Hint) {
+        let callouts = presentedCallouts.filter { $0.text == hint.text }
+        callouts.forEach { $0.dismiss() }
+    }
+    
+    
+    // MARK: - CalloutViewDelegate
+    
+    open func calloutViewDidDismiss(_ tipView: CalloutView) {
+        presentedCallouts = presentedCallouts.filter { $0.text != tipView.text }
+    }
+}
+
+
+// MARK: - Internal Functions
+
+extension CalloutHintPresenter {
+    
+    func createCallout(for hint: Hint) -> CalloutView {
+        let callout = CalloutView(text: hint.text, preferences: CalloutView.globalPreferences, delegate: self)
+        presentedCallouts.append(callout)
+        return callout
+    }
 }