+++ TuttiExample/Options/TableViewOption.swift
@@ -12,6 +12,7 @@ enum TableViewOption {
     case
     hint,
     deferredHint,
+    dismissCurrentHint,
     reset,
     tutorial,
     localizedTutorial,
@@ -27,6 +28,7 @@ extension TableViewOption {
         switch self {
         case .hint: return "Show a standard hint"
         case .deferredHint: return "Show a deferred hint"
+        case .dismissCurrentHint: return "Dismiss current hint"
         
         case .tutorial: return "Show a standard tutorial"
         case .localizedTutorial: return "Show a localized tutorial"
@@ -40,6 +42,7 @@ extension TableViewOption {
         switch self {
         case .hint: return "Hints are simple messages that are intended to be displayed once. They are perfect for quick, non-intrusive onboarding."
         case .deferredHint: return "Deferred hints are displayed after a custom number of presentation attempts. They are perfect for when you want to wait a while with displaying a hint."
+        case .dismissCurrentHint: return "This will try to find a matching hint view and dismiss it from the screen. You should do this on viewWillDisappear(...)."
         
         case .tutorial: return "Tutorials are single or multi page guides that are intended to display more information than a quick hint. They are perfect for modal screens or multi page guides."
         case .localizedTutorial: return "Localized tutorials automatically generate themselved by using available language keys in the current app. This saves you a lot of time and hassle."